update subtotal field

+++ granadatiles_project/apps/cart/serializers.py
@@ -11,7 +11,7 @@ class TileSerializer(BaseCatalogSerializer):
     size = serializers.CharField()
     image = serializers.CharField()
     colors = TileColorSerializer(many=True)
-    subtotal = serializers.IntegerField()
+    subtotal = serializers.FloatField()
 
 
 class TileOrdersSerializer(BaseSerializer):
